+++ Runtime/Inputs/InputAxisVector.cs
@@ -1,10 +1,10 @@
 ﻿using UnityEngine;
 using UnityEngine.InputSystem;
 
-namespace DapperDino.DapperTools.Components.Inputs
+namespace DapperDino.DapperTools.Inputs
 {
-    [CreateAssetMenu(fileName = "New Input Axis", menuName = "Inputs/Input Axis")]
-    public class InputAxis : ScriptableObject
+    [CreateAssetMenu(fileName = "New Input Axis Vector", menuName = "Inputs/Input Axis Vector")]
+    public class InputAxisVector : ScriptableObject
     {
         [SerializeField] private InputActionReference inputAction = null;