+++ Runtime/Inputs/InputReceiver.cs
@@ -1,13 +1,13 @@
-﻿using Assets.Dapper_Tools.Runtime.Components.Inputs;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using UnityEngine;
 
-namespace DapperDino.DapperTools.Components.Inputs
+namespace DapperDino.DapperTools.Inputs
 {
     public class InputReceiver : MonoBehaviour
     {
         [SerializeField] private int priority = 0;
-        [SerializeField] private List<InputAxisResponse> axisResponses = new List<InputAxisResponse>();
+        [SerializeField] private List<InputAxisFloatResponse> axisFloatResponses = new List<InputAxisFloatResponse>();
+        [SerializeField] private List<InputAxisVectorResponse> axisVectorResponses = new List<InputAxisVectorResponse>();
         [SerializeField] private List<InputActionResponse> actionResponses = new List<InputActionResponse>();
 
         public int Priority => priority;
@@ -17,10 +17,16 @@ namespace DapperDino.DapperTools.Components.Inputs
 
         public InputData ProcessInput(InputData inputData)
         {
-            // Raise an event with the axis values for all of the axis responses
-            foreach (var axisResponse in axisResponses)
+            // Raise an event with the axis values for all of the float axis responses
+            foreach (var axisResponse in axisFloatResponses)
             {
-                axisResponse.OnAction.Invoke(inputData.GetAxis(axisResponse.Axis));
+                axisResponse.OnAction.Invoke(inputData.GetAxisFloat(axisResponse.Axis));
+            }
+
+            // Raise an event with the axis values for all of the vector axis responses
+            foreach (var axisResponse in axisVectorResponses)
+            {
+                axisResponse.OnAction.Invoke(inputData.GetAxisVector(axisResponse.Axis));
             }
 
             // Raise an event for each of the action responses that had their action triggered this frame