+++ Runtime/SavingLoading/ISaveable.cs
@@ -1,0 +1,8 @@
+﻿namespace DapperDino.DapperTools.SavingLoading
+{
+    public interface ISaveable
+    {
+        object CaptureState();
+        void RestoreState(object state);
+    }
+}